Adds a standalone project to build customized IREE runtime libraries.
Just starting in iree-samples for now since this will always be a special way to build the runtime for niche uses, and keeping it isolated allows it to be forked/adapted at need.

This should help people looking to build language bindings and other systematic integrations, and it gets us out of the business of having to support all of the ways the runtime can be used (and abused) while lighting the path that we recommend. We likely need more discussion to make this a real thing but it is serviceable as-is.

Shared library support depends on https://github.com/iree-org/iree/pull/11723