[transform_dialect][README] Add a few "beware" comments
1.
The set of functions defined in scripts.sh expects the iree tools to
support debugging capabilities (like saving temporary files) and this
means they need to be built with assertions enabled.
Document this as one of the requirement.

2.
The output of `benchmark-transform-run-iree-compile` is a `.vmfb` file.
When building for CUDA, we won't notice it since the PTX part is dumped
last and is readable as text, but for `llvm-cpu` you get the elf binary
which is not that useful.
Add a comment about that.

NFC