[transform_dialect] Explicitly set the value for the `jit`
With iree now using by default the jit mode for CUDA workloads,
the script would choke with the interpreter mode because both the
jit mode and interpreter mode would be set for the compiler.
More precisely, the compilation would fail with:
"error: option clash in transform dialect lowering config: the filename
cannot be provided when the jit optio n is set"

To avoid this problem, always specify the jit mode. That way, we're
not affected by whatever is the default jit mode in the compiler.