Extract standalone variant of cuDNN dialect and lowering
Base set of cuDNN ops (with too rigid syntax!) and initial lowering to LLVM
dialect to construct the operation graph descriptor (given handle). The
conversion is keyed on the `build_and_exec_graph` op to create the build
function. Different lowerings depending on backend/runtime is
possible/expected. For end-to-end, finding the execution plan, allocating
workspace, data marshalling and execution is required. The dialect needs to be
expanded to target cuDNN backend API more fully and is definitely evolving
still.