Add first fp16 benchmarks.
This adds support to specify the data type when generating artifacts
for the benchmarking pipeline through PyTorch, and adds three benchmarks
that worked out of the box:

* ResNet50
* EfficientNet_V2
* BertLarge

Dumping FP16 HLO requires a machine with CUDA, as PyTorch needs to trace
the graph and is lacking support for most operations on CPU.

This required a bit of fiddling with torch-mlir and torch versions, as
the initial scripts pulled nightly torch installers without CUDA
enabled.

On machines without a GPU, the model import can still be executed as
before, and the import of models that require a GPU will be skipped.