Add a python script to help with correctness verification
Usage info can be found in the README. There are still a few todos to
streamline some of the more common cases (e.g. testing matmul just based
on shape info) but this is a starting point for people who just want to
be able to verify the IR they are already testing.